requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.11
google-generativeai>=0.4.1
serpapi==0.1.5
python-dotenv==1.0.0 